Always declare locals as mutable in derived impls
Values for fields are kept in local variables prior to construction in
the `From*` derivations. Before #177, these local variables would be
immutable, since nothing would ever modify them. However, now the
generated code will mutate the local variable to insert the result of
a `from_none` call to the child. Therefore, for simplicity the variables
are always emitted as mutable.

Fixes #183